🛠 Update of README.md…
And correction of an error. Should I create a wiki for the documentation ? Or a `/doc` directory and write everything in `.md` files ? 
I think the 1st solution is better because you won’t have to download the doc each time you wanna use the API or the docker image.